Extract printChildrent and printElement
